Implement traces issues detail pannel
Now we only display spans table but when clicking on a row it takes to
document traces tab with the spanId on the URL. In this PR we display in
place in the issues panel the details of the trace
