+++ apps/web/src/components/DetailsPannel/index.tsx
@@ -4,9 +4,11 @@ import { cn } from '@latitude-data/web-ui/utils'
 const DetailsPanel = ({
   ref,
   children,
+  bordered,
   className,
 }: {
   children: ReactNode
+  bordered: boolean
   className?: string
   ref?: Ref<HTMLDivElement>
 }) => {
@@ -14,9 +16,10 @@ const DetailsPanel = ({
     <div
       ref={ref}
       className={cn(
-        'flex flex-col flex-grow min-h-0 bg-background overflow-x-auto',
-        'border border-border rounded-lg items-center relative',
+        'flex flex-col gap-y-6 flex-grow min-h-0 bg-background overflow-x-auto',
+        'rounded-lg items-center relative',
         className,
+        { 'border border-border': bordered },
       )}
     >
       {children}
@@ -28,19 +31,38 @@ DetailsPanel.displayName = 'DetailsPanel'
 
 DetailsPanel.Header = function DetailsPanelHeader({
   children,
+  space = 'normal',
 }: {
   children: ReactNode
+  space?: 'normal' | 'none'
 }) {
-  return <div className='pt-6 px-4 relative w-full'>{children}</div>
+  return (
+    <div
+      className={cn('relative w-full', {
+        'px-4 pt-4': space === 'normal',
+      })}
+    >
+      {children}
+    </div>
+  )
 }
 
 DetailsPanel.Body = function DetailsPanelBody({
   children,
+  space = 'normal',
 }: {
   children: ReactNode
+  space?: 'normal' | 'none'
 }) {
   return (
-    <div className='w-full px-4 pb-5 mt-5 custom-scrollbar scrollable-indicator overflow-auto relative'>
+    <div
+      className={cn(
+        'w-full custom-scrollbar scrollable-indicator overflow-auto relative',
+        {
+          'px-4 pb-5': space === 'normal',
+        },
+      )}
+    >
       {children}
     </div>
   )
@@ -48,11 +70,19 @@ DetailsPanel.Body = function DetailsPanelBody({
 
 DetailsPanel.Footer = function DetailsPanelFooter({
   children,
+  space = 'normal',
 }: {
   children: ReactNode
+  space?: 'normal' | 'none'
 }) {
   return (
-    <div className='w-full bg-card rounded-b-lg p-4 mt-auto'>{children}</div>
+    <div
+      className={cn('w-full bg-card rounded-b-lg mt-auto', {
+        'p-4': space === 'normal',
+      })}
+    >
+      {children}
+    </div>
   )
 }
