+++ apps/web/src/components/TableResizableLayout/index.tsx
@@ -5,7 +5,7 @@ import {
 } from '@latitude-data/web-ui/atoms/Resizable'
 import { cn } from '@latitude-data/web-ui/utils'
 
-const MIN_RIGHT_PANE_WIDTH = 200
+const MIN_RIGHT_PANE_WIDTH = 560
 const RESIZE_HANDLES = ['w' as const]
 
 export function TableResizableLayout({
